refactor: Use werkzeug router for API routing
